Introduce proof-of-concept quality pod snapshotter
The idea of the Honeycomb agent is to allow users to configure
how the Honeytailer will parse the logs of some pod (e.g., specifying
parsers, sampling rates, and so on) on the basis of pod labels. So,
different parsers will parse the logs of some set of containers on the
basis of the label selectors specified by the user.

In order to facilitate this, it is necessary to be able to query the API
server for all nodes on a pod, subject to some label selector.

This commit will introduce a proof-of-concept quality version of this
code.

In particular, this code does NOT:

* Handle errors gracefully
* Even approach being prod-ready