Don't parse explicitly specified paths as Docker json-file logs
The implementation here is a bit messy, but will do for now. If given an
explicit path to watch (e.g. `/var/log/nginx/access.log`, we assume (for
now) that the log line is *not* in docker json-file format. Conversely, if
watching `/var/log/pods`, we assume that the log line *is* in json-file format.

In the future, we may want to extend the configuration to let folks specify
additional wrapper formats, e.g. perhaps
```
watchers:
  - dataset: mydataset
    paths:
    - /var/log/myservice.log
    wrapperFormat: syslog
```