Fix nil pointer dereference when state file open fails
This mystery turns out to be yet another episode of the unoriginal
crime series in which the serial culprit is always the same
(https://golang.org/doc/faq#nil_error), as is the unwitting victim (me).

If `StateRecorder` initialization fails, we log an error but continue
execution. All access to the `StateRecorder` is *supposed* to be guarded
by nil checks. However, an interface wrapping a nil pointer of a
concrete type does not compare equal to nil. Fix this by ensuring that
we return the nil *interface* from the `NewStateRecorder`, while
surpressing the urge to hurl all computers into the cold, bottomless
ocean.

Test plan: new unit tests