Add parser shortcut for nginx ingress logs
In Kubernetes, an _Ingress_ resource manages external access to
resources in a cluster. The details are kind of laden with Kubernetes
jargon, but the key point is that it's common to run NGINX as a sort of
cluster-wide proxy to the services that you run inside the cluster.

When run this way, nginx is configured to use a non-default logging
format. As a result, several users have reported trouble configuring
`honeycomb-kubernetes-agent` to parse nginx ingress logs. You can
technically do it by providing a custom log format in the agent's
configuration, but it's really verbose and opaque.

Make this easier by providing a shortcut: if you want to parse nginx
ingress logs, you can simply specify `parser: nginx-ingress` when
configuring the agent. I'll follow up with a docs patch as well.

Fixes #14.

Test plan: new unit tests.