[podtailer] support yet another log pattern
Looks like the log pattern changed again in 1.10. I am unable to find any documentation about this change in the changelog or elsewhere, but confirmed with minikube that the log directory pattern is indeed this:

`/var/log/pods/<podUID>/<containerName>/<instance>.log`

This is a change from pre-1.10 which was:

`/var/log/pods/<podUID>/<containerName>_instance.log`

In this diff we detect both and act accordingly. Not too thrilled about the approach in `determineFilterFunc` but this should get us working again in 1.10 clusters. With more cycles we can re-imagine the logging process to be version-aware or something.