Handle race between appearance of pod and log files on disk
Fixes https://github.com/honeycombio/honeycomb-kubernetes-agent/issues/25

When the K8s client signals a new pod to our watcher loop, the log files on disk are not necessarily in place yet. This can cause `determineLogPattern` to fail and return an empty pattern. As best as I can tell, we used to get away with this by falling back to a default pattern, but that's not always the best thing to do now that `determineLogPattern` is more complicated and has to address path differences between different versions of k8s. Instead, this adds a retry loop with backoff logic to the watcher loop. If `watcherForPod` fails, we'll retry a few times. In testing, this works, but there may be some edge cases out there where the log files do not appear after many seconds. If that happens, we'll at least have a log line to identify those cases and we can come up with a better strategy.