Add optional error dataset for parsing errors
This is similar to our implementation in the Agentless Integrations. Rather than log parsing errors locally (requiring debug mode and knowledge of kubectl-foo), allow an optional `ErrorDataset` in the config - either globally, or on a per-watcher basis. Parsing errors will get logged to this dataset. It's also possible to log errors to the same dataset as the watcher, but it's probably not a good idea so you have to be explicit about it.