refactor pathwatcher to accept a pattern function
Using the structure of the files on disk is racey and problematic. What if the pod doesn't write any lines for an hour after it comes online? We can't wait to start the watcher that long (since we read off the channel in order, this would block other watchers from starting) and we don't want to fail the watcher and forget about the pod entirely.

Rather than determine the log pattern at watcher start time, do it after the watcher starts by passing the pattern function to the watcher.