[versioning] switch to semantic versioning, publish versions on tag builds only.
The current versioning strategy - publish every master build as `head` - is limiting and makes it hard for users to confidently deploy a stable agent. It also makes it tricky to debug user reports, as everyone is using "head" but that doesn't really mean anything. If we ever want to publish a changelog or start building deployments for Helm or other tools, we'll need to sort out our versioning anyway. This moves us to use an explicit version, defined in version.txt and injected into the binary as part of the build process.

I started at 1.2.0, as we already had some files referencing "1.1".