[vendor] pull in k8s client v11, refactor as necessary
Started down this path after trying to add cluster name to the metadata included in k8s log events, and realizing we were on version 2.0.0 of the client code while the latest version is 11.0.0. Turns out, cluster_name still isn't passed in in the latest version of the k8s client, but at least this refactoring is done.