Support newer log path pattern, bump to 1.3.1
Our agent is broken on K8S cluster versions published after this change: https://github.com/kubernetes/kubernetes/pull/74441

Adding yet another log path bandaid to the list of known patterns. It works, but I'd like to find a way to stop doing this. Unfortunately the current K8S API doesn't provide a way to get the log *path* for a pod, just the log. We could start streaming k8s logs from the API, but that's a significant amount of work and we don't have time to do that right now. Here's hoping the log pattern doesn't change again in the next 6 months.