processors: Add a pair of simple filter processors
This filters events based on exact matches of key:values, in the form of two
basic processors.

keep_event:
  Matches a single key to one of a series of values, and
  keeps only events matching this value.

drop_event:
  Matches a single key to one of a series of values, and drops all events matching
  this value

These are intended as processor filters from things like the `nginx-ingress`
which contains data from one or more datasets, which should be split out based
on some pre-configured conditions.

Typical usecase was to filter on the k8s.service or a specific hostname.

It only performs exact string matches, as it is not intended to be a generic
filter or routing implementation.