processors/route_event: A simple value based event router (#53)
This lets one split events in kubernetes into different datasets based on a
simple route matching logic.

This is mostly useful for services that serve multiple different backends, like
load balancers, ingresses and databases. It was specifically made to handle
logs out of `nginx-ingress` to be able to split them based on the service it
waas routing data to.