Deflake flaky tests (#264)
## Which problem is this PR solving?

The polling strategy of the podtailer has a re-poll time of 1 second; the best guess I had for the flaky tests only in CI was that on a CPU-limited CI system, the 1-second sleep in the test was too fast for the 1-second retry time of the podtailer. On the laptop with many cores, there's plenty of CPU around to run the tailers in a separate thread. 

This matters for this test because it's rotating logs and then closing them quickly; under normal circumstances (even under small CPU allocations), the tailer should have plenty of time to catch up to a rotated log.

## Short description of the changes

- Lengthen the sleep time in the flaky test
- Lengthen the wait time for Bolt to acquire a file lock

Fixes #263 
Fixes #225