move publishing image to separate job
This separates the act of publishing the image to Docker Hub (requires
secrets) from the act of building and testing (does not require
secrets). The command to publish is much simplified now that the job is
only run on version-tagged commits.

Secrets needed to publish are provided through CircleCI Context as
recommended by Circle.

Other changes:

* reused filter definitions common in our other CI configs
* declared the version of Go to build with in a single named executor